requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
Django==5.1.4
django-anymail==12.0
filelock==3.16.1
frozenlist==1.5.0
future==1.0.0
idna==3.10
imageio==2.36.1
maxminddb==2.6.3
multidict==6.1.0
mysqlclient==2.2.6
numpy==2.2.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
python-dotenv==1.0.1
python-http-client==3.3.7
ratelim==0.1.6
razorpay==1.4.2
requests==2.32.3
sendgrid==6.11.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
wheel==0.45.1
yarl==1.18.3
